pyproject.toml
[build-system]
requires = ["maturin>=1.12,<2.0"]
build-backend = "maturin"

[project]
name = "ravif_py"
description = "Fast AVIF encoder for Python based on ravif (Rust)"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "numpy>=1.24.4",
    "pillow>=10.4.0",
]
[tool.maturin]
module-name = "ravif_py._ravif_py"